Dockerfile: Configure user id at build
Configure user id, group id and user name at docker build time. Move the
west-workspace into the user-home directory. The user needs to be the
owner of the west-workspace. It is more natural for it to lay in the
user's home directory. Furtheradvantage is, that if the UID needs to be
changed later as part of the docker startup all files in the home
directory are automatically adjusted.