Dockerfile: rearrange apt package installation for overview/readability
Use long-form command line parameter where possible for readability.
When installing apt packages it is not necessary to clean after, since
according to the
[docs](https://docs.docker.com/build/building/best-practices/#apt-get)
it is done automatically in official debian and ubuntu images. But it is
necessary to remove the cache in `/var/lib/apt/lists/*`. This makes it
necessary to update again before using apt in another layer.
Use a single package per line. This makes it easier to analyze diffs, if
a package was added or removed. All packages that are needed by zephyr
according to the
[documentation](https://docs.zephyrproject.org/latest/develop/getting_started/index.html#install-dependencies).
This makes it easier to see which packages where added by us and which
are necessary by zephyr.